backend/pyproject.toml
[project]
name = "rag-knowledge-base-backend"
version = "0.1.0"
description = "FastAPI backend for a RAG-powered knowledge base."
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.13.0",
  "chromadb>=0.5.0",
  "fastapi>=0.115.0",
  "openai>=1.50.0",
  "psycopg[binary]>=3.2.0",
  "pydantic-settings>=2.5.0",
  "pymupdf>=1.24.0",
  "python-multipart>=0.0.9",
  "redis>=5.0.0",
  "rq>=1.16.0",
  "sqlalchemy>=2.0.0",
  "tiktoken>=0.7.0",
  "uvicorn[standard]>=0.30.0"
]

[project.optional-dependencies]
dev = [
  "httpx>=0.27.0",
  "pytest>=8.0.0",
  "ruff>=0.6.0"
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
testpaths = ["tests"]
